hook処理の追加とリファクタリング
